backend/requirements.txt
fastapi
uvicorn
SPARQLWrapper
psycopg2
owlready2
python-dotenv
python-multipart